Move helper methods to helper file
